+++ CostFunctions/CostFunction.cpp
@@ -2,19 +2,19 @@
 #include <cmath>
 #include <numeric>
 
-CostFunction::CostFunction(double regularizationLambda) : regularizationLambda(regularizationLambda) {}
+CostFunction::CostFunction(float regularizationLambda) : regularizationLambda(regularizationLambda) {}
 
-double CostFunction::getActivationDerivative(double activation, double expected) const {
+float CostFunction::getActivationDerivative(float activation, float expected) const {
     return activation - expected;
 }
 
-double CostFunction::getRegularizationCost(std::vector<double> weights) const {
-    double sum = std::accumulate(weights.begin(), weights.end(), 0.0, [](double s, double w) {
+float CostFunction::getRegularizationCost(std::vector<float> weights) const {
+    float sum = std::accumulate(weights.begin(), weights.end(), 0.0f, [](float s, float w) {
         return s + std::pow(w, 2);
     });
     return (regularizationLambda * sum) / (2 * totalWeightCount);
 }
 
-double CostFunction::getRegularizationDerivative(double weight) const {
+float CostFunction::getRegularizationDerivative(float weight) const {
     return (regularizationLambda * weight) / totalWeightCount;
 }
