+++ NeuralNetworks/Neuron.h
@@ -9,28 +9,28 @@ class Layer;
 class CostFunction;
 
 struct Weight {
-    double weight = 0;
-    double gradient = 0;
+    float weight = 0;
+    float gradient = 0;
 };
 
 class Neuron {
 public:
-    double value = 0.0;
+    float value = 0.0;
 private:
-    double bias = 0.0;
-    double biasGradient = 0.0;
+    float bias = 0.0;
+    float biasGradient = 0.0;
 
     std::vector<Weight> weights;
-    double activationGradient = 0.0;
-    double preTransformedValue = 0.0;
+    float activationGradient = 0.0;
+    float preTransformedValue = 0.0;
     
     const ActivationFunction& activationFunction;
     const CostFunction& costFunction;
 public:
     Neuron(const ActivationFunction& activationFunction, const CostFunction& costFunction, const Layer* previousLayer, const Layer* nextLayer);
     void calculate(const Layer& previousLayer);
-    void addActivationGradient(double gradient);
-    void backpropagate(Layer& previousLayer, std::optional<double> expectedValue = std::nullopt);
-    void update(double batchSize, double learningRate);
+    void addActivationGradient(float gradient);
+    void backpropagate(Layer& previousLayer, std::optional<float> expectedValue = std::nullopt);
+    void update(int32_t batchSize, float learningRate);
     size_t getWeightCount();
 };
